pyproject.toml
[tool.poetry]
name = "flask-project"
version = "0.1.0"
description = ""
authors = ["maksik <mzagrebam@gmail.com>"]
readme = "README.md"
packages = [{include = "flask_project"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
python-dotenv = "^1.0.0"
flask-sqlalchemy = "^3.0.3"
flask-login = "^0.6.2"
wtforms = "^3.0.1"
email-validator = "^1.3.1"
flask-wtf = "^1.1.1"
flask-admin = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
